backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.0
python-dotenv==1.0.0
psycopg2-binary==2.9.9
django-environ==0.11.2
drf-yasg==1.21.7
Pillow==10.0.1
celery==5.3.4
redis==5.0.1
whitenoise==6.6.0
gunicorn==21.2.0
